separate unconstrained from struct
