+++ src/index.ts
@@ -22,12 +22,8 @@ export type {
   FlexibleProvablePure,
   InferProvable,
 } from './lib/provable-types/struct.js';
-export {
-  provable,
-  provablePure,
-  Struct,
-  Unconstrained,
-} from './lib/provable-types/struct.js';
+export { provable, provablePure, Struct } from './lib/provable-types/struct.js';
+export { Unconstrained } from './lib/provable-types/unconstrained.js';
 export { Provable } from './lib/provable.js';
 export { Circuit, Keypair, public_, circuitMain } from './lib/circuit.js';
 export { UInt32, UInt64, Int64, Sign, UInt8 } from './lib/int.js';
