+++ src/lib/provable-types/merkle-list.ts
@@ -1,9 +1,10 @@
 import { Bool, Field } from '../core.js';
 import { Provable } from '../provable.js';
-import { Struct, Unconstrained } from './struct.js';
+import { Struct } from './struct.js';
 import { assert } from '../gadgets/common.js';
 import { provableFromClass } from './provable-derivers.js';
 import { Poseidon, packToFields, ProvableHashable } from '../hash.js';
+import { Unconstrained } from './unconstrained.js';
 
 export {
   MerkleListBase,
