+++ src/lib/provable-types/packed.ts
@@ -1,5 +1,6 @@
 import { provableFromClass } from './provable-derivers.js';
-import { HashInput, ProvableExtended, Unconstrained } from './struct.js';
+import { HashInput, ProvableExtended } from './struct.js';
+import { Unconstrained } from './unconstrained.js';
 import { Field } from '../field.js';
 import { assert } from '../gadgets/common.js';
 import { Poseidon, ProvableHashable, packToFields } from '../hash.js';
